change folder structure to be same as cookiecutter template
