--- js/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-<!-- <START NEW CHANGELOG ENTRY> -->
-
-<!-- <END NEW CHANGELOG ENTRY> -->
