--- js/MANIFEST.in
@@ -1,24 +1,0 @@
-include LICENSE
-include *.md
-include pyproject.toml
-
-include package.json
-include install.json
-include ts*.json
-include yarn.lock
-
-graft xellgrid2/labextension
-
-# Javascript files
-graft src
-graft style
-prune **/node_modules
-prune lib
-prune binder
-
-# Patterns to exclude from any directory
-global-exclude *~
-global-exclude *.pyc
-global-exclude *.pyo
-global-exclude .git
-global-exclude .ipynb_checkpoints
