--- js/xellgrid2/__init__.py
@@ -1,17 +1,0 @@
-
-import json
-from pathlib import Path
-
-from ._version import __version__
-
-HERE = Path(__file__).parent.resolve()
-
-with (HERE / "labextension" / "package.json").open() as fid:
-    data = json.load(fid)
-
-def _jupyter_labextension_paths():
-    return [{
-        "src": "labextension",
-        "dest": data["name"]
-    }]
-
