+++ src/xellgrid.booleanfilter.js
