+++ src/xellgrid.sliderfilter.js
