+++ src/xellgrid.textfilter.js
