+++ src/xellgrid.widget.js
